honksquad: test: split oversized RussStation integration test fixtures by scenario (#868)
Splits the five fork integration test files flagged in the Lens E tests
audit into scenario-grouped partial-class files, following the existing
`Foo.cs` + `Foo.Scenario.cs` convention (e.g. BuckleTest / MindTests).
Each fixture stays a single NUnit `[TestFixture]`, so the shared
`[TestPrototypes]` blocks and helpers are not duplicated and test
discovery/behavior is unchanged.

- GasReactionTest -> + ProtoNitrate (proto-nitrate gas chain),
  + AdjustEnergy (ReactionHelper energy/temperature helper)
- SurgerySystemTest -> + Tools (tool matching/tier), + Modifiers
  (surface/drape/difficulty duration); main keeps prototype/preset validation
- EscalatedGrabTest -> + Movement (speed/strip effects),
  + Combat (choke/damage/resist); main keeps stage state queries
- TraitPointBuyTest -> + TagExclusion (PR #381); main keeps the point
  budget / category caps (PR #376)
- SkillchipGrantTest -> + Constraints (capacity/duplicate/category);
  main keeps the grant lifecycle

No test logic changed; all 84 tests across the five fixtures are preserved.
