+++ Content.IntegrationTests/Tests/RussStation/Atmos/GasReactionTest.cs
@@ -14,11 +14,16 @@ namespace Content.IntegrationTests.Tests.RussStation.Atmos;
 /// <c>Content.Server/RussStation/Atmos/Reactions</c>. Each reaction is
 /// exercised with a happy-path input and at least one no-reaction input,
 /// plus branch-specific cases where a reaction has more than one early
-/// return (BZ decomposition). Also covers
-/// <see cref="ReactionHelper.AdjustEnergy"/> directly because every reaction
-/// depends on it.
+/// return (BZ decomposition).
 /// </summary>
-public sealed class GasReactionTest : GameTest
+/// <remarks>
+/// This partial holds the <see cref="TestReaction{T}"/> helper and the simpler
+/// single-reaction cases. The proto-nitrate gas chain lives in
+/// <c>GasReactionTest.ProtoNitrate.cs</c>, and the
+/// <see cref="ReactionHelper.AdjustEnergy"/> cases (which every reaction depends on)
+/// live in <c>GasReactionTest.AdjustEnergy.cs</c>.
+/// </remarks>
+public sealed partial class GasReactionTest : GameTest
 {
     private const double Tol = 1e-3;
 
@@ -346,248 +351,4 @@ public sealed class GasReactionTest : GameTest
                 Assert.That(mix.GetMoles(Gas.Nitrogen), Is.Zero);
             });
         });
-
-    // ---- ProtoNitrateFormation ---------------------------------------------
-
-    [Test]
-    public Task ProtoNitrateFormation_PluoxiumAndHydrogenHot_ProducesProtoNitrate() => TestReaction<ProtoNitrateFormationReaction>(
-        mix =>
-        {
-            mix.Temperature = 6000f;
-            mix.AdjustMoles(Gas.Pluoxium, 10f);
-            mix.AdjustMoles(Gas.Hydrogen, 20f);
-        },
-        (result, mix, _) =>
-        {
-            // heatEff = min(6000*0.005=30, min(10/0.2=50, 20/2=10)) = 10
-            // consumes 2 pluoxium + 20 hydrogen, produces 22 proto-nitrate.
-            // The reaction adds positive energy, but the product has much
-            // more heat capacity than the reactants so the final temperature
-            // is a function of both; only assert moles here.
-            Assert.Multiple(() =>
-            {
-                Assert.That(result, Is.EqualTo(ReactionResult.Reacting));
-                Assert.That(mix.GetMoles(Gas.Pluoxium), Is.EqualTo(8f).Within(Tol));
-                Assert.That(mix.GetMoles(Gas.Hydrogen), Is.Zero.Within(Tol));
-                Assert.That(mix.GetMoles(Gas.ProtoNitrate), Is.EqualTo(22f).Within(Tol));
-            });
-        });
-
-    [Test]
-    public Task ProtoNitrateFormation_NoPluoxium_DoesNotReact() => TestReaction<ProtoNitrateFormationReaction>(
-        mix =>
-        {
-            mix.Temperature = 6000f;
-            mix.AdjustMoles(Gas.Hydrogen, 20f);
-        },
-        (result, mix, _) =>
-        {
-            Assert.Multiple(() =>
-            {
-                Assert.That(result, Is.EqualTo(ReactionResult.NoReaction));
-                Assert.That(mix.GetMoles(Gas.ProtoNitrate), Is.Zero);
-            });
-        });
-
-    // ---- ProtoNitrateHydrogen ----------------------------------------------
-
-    [Test]
-    public Task ProtoNitrateHydrogen_WithHydrogen_ConvertsToMoreProtoNitrateAndCools() => TestReaction<ProtoNitrateHydrogenReaction>(
-        mix =>
-        {
-            mix.Temperature = 400f;
-            mix.AdjustMoles(Gas.Hydrogen, 10f);
-            mix.AdjustMoles(Gas.ProtoNitrate, 10f);
-        },
-        (result, mix, _) =>
-        {
-            // produced = min(5, min(10, 10)) = 5
-            // consumes 5 H2, adds 2.5 proto-nitrate
-            Assert.Multiple(() =>
-            {
-                Assert.That(result, Is.EqualTo(ReactionResult.Reacting));
-                Assert.That(mix.GetMoles(Gas.Hydrogen), Is.EqualTo(5f).Within(Tol));
-                Assert.That(mix.GetMoles(Gas.ProtoNitrate), Is.EqualTo(12.5f).Within(Tol));
-                Assert.That(mix.Temperature, Is.LessThan(400f));
-            });
-        });
-
-    [Test]
-    public Task ProtoNitrateHydrogen_NoHydrogen_DoesNotReact() => TestReaction<ProtoNitrateHydrogenReaction>(
-        mix =>
-        {
-            mix.Temperature = 400f;
-            mix.AdjustMoles(Gas.ProtoNitrate, 10f);
-        },
-        (result, mix, _) =>
-        {
-            Assert.Multiple(() =>
-            {
-                Assert.That(result, Is.EqualTo(ReactionResult.NoReaction));
-                Assert.That(mix.GetMoles(Gas.ProtoNitrate), Is.EqualTo(10f).Within(Tol));
-            });
-        });
-
-    // ---- ProtoNitrateTritium -----------------------------------------------
-
-    [Test]
-    public Task ProtoNitrateTritium_WithTritium_ConvertsToHydrogenAndHeats() => TestReaction<ProtoNitrateTritiumReaction>(
-        mix =>
-        {
-            mix.Temperature = 300f;
-            mix.AdjustMoles(Gas.Tritium, 20f);
-            mix.AdjustMoles(Gas.ProtoNitrate, 10f);
-        },
-        (result, mix, _) =>
-        {
-            // first = 300/34 * (20*10)/(20+10*10) = 8.824 * 1.667 = 14.706
-            // second = min(20, 10/0.01=1000) = 20
-            // produced = min(14.706, 20) = 14.706
-            const float produced = 300f / 34f * (20f * 10f) / (20f + 10f * 10f);
-            Assert.Multiple(() =>
-            {
-                Assert.That(result, Is.EqualTo(ReactionResult.Reacting));
-                Assert.That(mix.GetMoles(Gas.Tritium), Is.EqualTo(20f - produced).Within(Tol));
-                Assert.That(mix.GetMoles(Gas.ProtoNitrate), Is.EqualTo(10f - produced * 0.01f).Within(Tol));
-                Assert.That(mix.GetMoles(Gas.Hydrogen), Is.EqualTo(produced).Within(Tol));
-                Assert.That(mix.Temperature, Is.GreaterThan(300f));
-            });
-        });
-
-    [Test]
-    public Task ProtoNitrateTritium_NoTritium_DoesNotReact() => TestReaction<ProtoNitrateTritiumReaction>(
-        mix =>
-        {
-            mix.Temperature = 300f;
-            mix.AdjustMoles(Gas.ProtoNitrate, 10f);
-        },
-        (result, mix, _) =>
-        {
-            Assert.Multiple(() =>
-            {
-                Assert.That(result, Is.EqualTo(ReactionResult.NoReaction));
-                Assert.That(mix.GetMoles(Gas.Hydrogen), Is.Zero);
-            });
-        });
-
-    // ---- ProtoNitrateBZ ----------------------------------------------------
-
-    [Test]
-    public Task ProtoNitrateBZ_WithBZ_DecomposesIntoN2HeAndPlasma() => TestReaction<ProtoNitrateBZReaction>(
-        mix =>
-        {
-            mix.Temperature = 270f;
-            mix.AdjustMoles(Gas.BZ, 10f);
-            mix.AdjustMoles(Gas.ProtoNitrate, 10f);
-        },
-        (result, mix, _) =>
-        {
-            // consumed = min(270/2240 * 100/20, min(10,10)) = min(0.6027, 10) = 0.6027
-            const float consumed = 270f / 2240f * (10f * 10f) / (10f + 10f);
-            Assert.Multiple(() =>
-            {
-                Assert.That(result, Is.EqualTo(ReactionResult.Reacting));
-                Assert.That(mix.GetMoles(Gas.BZ), Is.EqualTo(10f - consumed).Within(Tol));
-                Assert.That(mix.GetMoles(Gas.Nitrogen), Is.EqualTo(consumed * 0.4f).Within(Tol));
-                Assert.That(mix.GetMoles(Gas.Helium), Is.EqualTo(consumed * 1.6f).Within(Tol));
-                Assert.That(mix.GetMoles(Gas.Plasma), Is.EqualTo(consumed * 0.8f).Within(Tol));
-                Assert.That(mix.Temperature, Is.GreaterThan(270f));
-            });
-        });
-
-    [Test]
-    public Task ProtoNitrateBZ_NoBZ_DoesNotReact() => TestReaction<ProtoNitrateBZReaction>(
-        mix =>
-        {
-            mix.Temperature = 270f;
-            mix.AdjustMoles(Gas.ProtoNitrate, 10f);
-        },
-        (result, mix, _) =>
-        {
-            Assert.Multiple(() =>
-            {
-                Assert.That(result, Is.EqualTo(ReactionResult.NoReaction));
-                Assert.That(mix.GetMoles(Gas.ProtoNitrate), Is.EqualTo(10f).Within(Tol));
-            });
-        });
-
-    // ---- ReactionHelper.AdjustEnergy --------------------------------------
-
-    [Test]
-    public async Task AdjustEnergy_PositiveEnergy_HeatsMixture()
-    {
-        var server = Server;
-        var atmos = server.ResolveDependency<IEntitySystemManager>().GetEntitySystem<AtmosphereSystem>();
-
-        await server.WaitAssertion(() =>
-        {
-            var mix = new GasMixture(Atmospherics.CellVolume) { Temperature = 300f };
-            mix.AdjustMoles(Gas.Nitrogen, 100f);
-            var oldCap = atmos.GetHeatCapacity(mix, true);
-            ReactionHelper.AdjustEnergy(mix, atmos, oldCap, 100_000f, 1f);
-            Assert.That(mix.Temperature, Is.GreaterThan(300f));
-        });
-    }
-
-    [Test]
-    public async Task AdjustEnergy_NegativeEnergy_CoolsMixture()
-    {
-        var server = Server;
-        var atmos = server.ResolveDependency<IEntitySystemManager>().GetEntitySystem<AtmosphereSystem>();
-
-        await server.WaitAssertion(() =>
-        {
-            var mix = new GasMixture(Atmospherics.CellVolume) { Temperature = 300f };
-            mix.AdjustMoles(Gas.Nitrogen, 100f);
-            var oldCap = atmos.GetHeatCapacity(mix, true);
-            ReactionHelper.AdjustEnergy(mix, atmos, oldCap, -50_000f, 1f);
-            Assert.That(mix.Temperature, Is.LessThan(300f));
-        });
-    }
-
-    [Test]
-    public async Task AdjustEnergy_ExtremeNegativeEnergy_ClampsToTCMB()
-    {
-        var server = Server;
-        var atmos = server.ResolveDependency<IEntitySystemManager>().GetEntitySystem<AtmosphereSystem>();
-
-        await server.WaitAssertion(() =>
-        {
-            var mix = new GasMixture(Atmospherics.CellVolume) { Temperature = 300f };
-            mix.AdjustMoles(Gas.Nitrogen, 100f);
-            var oldCap = atmos.GetHeatCapacity(mix, true);
-            ReactionHelper.AdjustEnergy(mix, atmos, oldCap, -1_000_000_000f, 1f);
-            Assert.That(mix.Temperature, Is.EqualTo(Atmospherics.TCMB).Within(Tol));
-        });
-    }
-
-    [Test]
-    public async Task AdjustEnergy_HeatScaleHalves_HalvesEnergyApplied()
-    {
-        var server = Server;
-        var atmos = server.ResolveDependency<IEntitySystemManager>().GetEntitySystem<AtmosphereSystem>();
-
-        await server.WaitAssertion(() =>
-        {
-            var mix1 = new GasMixture(Atmospherics.CellVolume) { Temperature = 300f };
-            mix1.AdjustMoles(Gas.Nitrogen, 100f);
-            var cap1 = atmos.GetHeatCapacity(mix1, true);
-            ReactionHelper.AdjustEnergy(mix1, atmos, cap1, 100_000f, 1f);
-
-            var mix2 = new GasMixture(Atmospherics.CellVolume) { Temperature = 300f };
-            mix2.AdjustMoles(Gas.Nitrogen, 100f);
-            var cap2 = atmos.GetHeatCapacity(mix2, true);
-            ReactionHelper.AdjustEnergy(mix2, atmos, cap2, 100_000f, 2f);
-
-            // heatScale divides energy, so scale=2 should yield a smaller delta above 300K
-            var delta1 = mix1.Temperature - 300f;
-            var delta2 = mix2.Temperature - 300f;
-            Assert.Multiple(() =>
-            {
-                Assert.That(delta1, Is.GreaterThan(0f));
-                Assert.That(delta2, Is.GreaterThan(0f));
-                Assert.That(delta2, Is.EqualTo(delta1 / 2f).Within(0.1));
-            });
-        });
-    }
 }
