+++ Content.IntegrationTests/Tests/RussStation/Surgery/SurgerySystemTest.cs
@@ -1,18 +1,20 @@
 using Content.IntegrationTests.Fixtures;
-using Content.Server.RussStation.Surgery;
-using Content.Shared.Buckle;
-using Content.Shared.Buckle.Components;
 using Content.Shared.RussStation.Surgery;
-using Content.Shared.RussStation.Surgery.Components;
 using Content.Shared.RussStation.Surgery.Systems;
 using Content.Shared.Tools;
-using Robust.Shared.GameObjects;
 using Robust.Shared.Prototypes;
 
 namespace Content.IntegrationTests.Tests.RussStation.Surgery;
 
+/// <summary>
+/// Tests for <see cref="SharedSurgerySystem"/>. The procedure prototype and step-preset
+/// validation scenarios live here; tool matching/tier scenarios live in
+/// <c>SurgerySystemTest.Tools.cs</c>, and surface/drape/difficulty duration scenarios live
+/// in <c>SurgerySystemTest.Modifiers.cs</c>. All partials share the <see cref="Prototypes"/>
+/// fixture below.
+/// </summary>
 [TestOf(typeof(SharedSurgerySystem))]
-public sealed class SurgerySystemTest : GameTest
+public sealed partial class SurgerySystemTest : GameTest
 {
     private const string TestProcedureId = "SurgeryTestProcedure";
     private const string TestProcedureMajorId = "SurgeryTestProcedureMajor";
@@ -207,180 +209,22 @@ public sealed class SurgerySystemTest : GameTest
     }
 
     /// <summary>
-    /// Verifies that ToolMatchesStep correctly matches tool qualities to step requirements.
+    /// Verifies that the major test procedure loads with the correct difficulty.
     /// </summary>
     [Test]
-    public async Task ToolMatchesStepTest()
+    public async Task ProcedureDifficultyLoadsTest()
     {
         var server = Server;
-
-        var entityManager = server.ResolveDependency<IEntityManager>();
         var protoManager = server.ResolveDependency<IPrototypeManager>();
-        var mapData = await Pair.CreateTestMap();
 
         await server.WaitAssertion(() =>
         {
-            var surgerySystem = entityManager.System<SharedSurgerySystem>();
-            protoManager.TryIndex<SurgeryProcedurePrototype>(TestProcedureId, out var proto);
+            protoManager.TryIndex<SurgeryProcedurePrototype>(TestProcedureId, out var standard);
+            Assert.That(standard!.Difficulty, Is.EqualTo(SurgeryDifficulty.Standard));
 
-            var scalpel = entityManager.SpawnEntity("SurgeryTestScalpel", mapData.GridCoords);
-            var retractor = entityManager.SpawnEntity("SurgeryTestRetractor", mapData.GridCoords);
-
-            // Scalpel matches step 0 (Slicing), not step 1 (Retracting)
-            Assert.That(surgerySystem.ToolMatchesStep(scalpel, proto!.Steps[0]), Is.True);
-            Assert.That(surgerySystem.ToolMatchesStep(scalpel, proto.Steps[1]), Is.False);
-
-            // Retractor matches step 1 (Retracting), not step 0 (Slicing)
-            Assert.That(surgerySystem.ToolMatchesStep(retractor, proto.Steps[0]), Is.False);
-            Assert.That(surgerySystem.ToolMatchesStep(retractor, proto.Steps[1]), Is.True);
+            protoManager.TryIndex<SurgeryProcedurePrototype>(TestProcedureMajorId, out var major);
+            Assert.That(major!.Difficulty, Is.EqualTo(SurgeryDifficulty.Major));
         });
-    }
-
-    /// <summary>
-    /// Verifies that a tool with the correct quality but without a tier tag
-    /// still matches the step. Tier tags affect duration, not step matching.
-    /// </summary>
-    [Test]
-    public async Task ToolMatchesStepIgnoresTierTagTest()
-    {
-        var server = Server;
-
-        var entityManager = server.ResolveDependency<IEntityManager>();
-        var protoManager = server.ResolveDependency<IPrototypeManager>();
-        var mapData = await Pair.CreateTestMap();
-
-        await server.WaitAssertion(() =>
-        {
-            var surgerySystem = entityManager.System<SharedSurgerySystem>();
-            protoManager.TryIndex<SurgeryProcedurePrototype>(TestProcedureId, out var proto);
-
-            // Entity has Slicing quality but no tier tag (improvised)
-            var nonTool = entityManager.SpawnEntity("SurgeryTestNonTool", mapData.GridCoords);
-
-            // ToolMatchesStep only checks the quality, so this matches
-            Assert.That(surgerySystem.ToolMatchesStep(nonTool, proto!.Steps[0]), Is.True);
-        });
-    }
-
-    /// <summary>
-    /// Verifies that IsCauteryTool identifies cautery tools correctly.
-    /// </summary>
-    [Test]
-    public async Task IsCauteryToolTest()
-    {
-        var server = Server;
-
-        var entityManager = server.ResolveDependency<IEntityManager>();
-        var mapData = await Pair.CreateTestMap();
-
-        await server.WaitAssertion(() =>
-        {
-            var surgerySystem = entityManager.System<SharedSurgerySystem>();
-
-            var cautery = entityManager.SpawnEntity("SurgeryTestCautery", mapData.GridCoords);
-            var scalpel = entityManager.SpawnEntity("SurgeryTestScalpel", mapData.GridCoords);
-
-            Assert.That(surgerySystem.IsCauteryTool(cautery), Is.True);
-            Assert.That(surgerySystem.IsCauteryTool(scalpel), Is.False);
-        });
-    }
-
-    /// <summary>
-    /// Verifies that GetSurfaceSpeedModifier returns 2.0 for unbuckled patients
-    /// and the configured modifier when buckled to a surgery surface.
-    /// </summary>
-    [Test]
-    public async Task SurfaceSpeedModifierTest()
-    {
-        var server = Server;
-
-        var entityManager = server.ResolveDependency<IEntityManager>();
-        var mapData = await Pair.CreateTestMap();
-
-        await server.WaitAssertion(() =>
-        {
-            var surgerySystem = entityManager.System<SharedSurgerySystem>();
-            var buckleSystem = entityManager.System<SharedBuckleSystem>();
-
-            var patient = entityManager.SpawnEntity("SurgeryTestPatient", mapData.GridCoords);
-            var table = entityManager.SpawnEntity("SurgeryTestOperatingTable", mapData.GridCoords);
-
-            // Unbuckled patient should return floor penalty of 2.0
-            Assert.That(surgerySystem.GetSurfaceSpeedModifier(patient), Is.EqualTo(2f));
-
-            // Buckle to operating table -> 1.0x modifier
-            var buckle = entityManager.GetComponent<BuckleComponent>(patient);
-            Assert.That(buckleSystem.TryBuckle(patient, patient, table, buckleComp: buckle), Is.True);
-            Assert.That(surgerySystem.GetSurfaceSpeedModifier(patient), Is.EqualTo(1f));
-        });
-    }
-
-    /// <summary>
-    /// Verifies that GetSurfaceSpeedModifier returns 2.0 when buckled to an entity
-    /// that has a Strap component but no SurgerySurface component (e.g. a regular chair).
-    /// </summary>
-    [Test]
-    public async Task NonSurgerySurfaceReturnsDefaultModifierTest()
-    {
-        var server = Server;
-
-        var entityManager = server.ResolveDependency<IEntityManager>();
-        var mapData = await Pair.CreateTestMap();
-
-        await server.WaitAssertion(() =>
-        {
-            var surgerySystem = entityManager.System<SharedSurgerySystem>();
-            var buckleSystem = entityManager.System<SharedBuckleSystem>();
-
-            var patient = entityManager.SpawnEntity("SurgeryTestPatient", mapData.GridCoords);
-
-            // Plain strap with no SurgerySurface component
-            var chair = entityManager.SpawnEntity("SurgeryTestChair", mapData.GridCoords);
-
-            var buckle = entityManager.GetComponent<BuckleComponent>(patient);
-            Assert.That(buckleSystem.TryBuckle(patient, patient, chair, buckleComp: buckle), Is.True);
-            Assert.That(surgerySystem.GetSurfaceSpeedModifier(patient), Is.EqualTo(2f));
-        });
-    }
-
-    /// <summary>
-    /// Verifies that GetDrapeSpeedModifier returns the correct multiplier.
-    /// No drape component = 1.0 (no penalty), default drape = 1.5 (bedsheet improvised penalty).
-    /// Surgical drape stamping (1.0x) is tested server-side via the draping interaction.
-    /// </summary>
-    [Test]
-    public async Task DrapeSpeedModifierTest()
-    {
-        var server = Server;
-
-        var entityManager = server.ResolveDependency<IEntityManager>();
-        var mapData = await Pair.CreateTestMap();
-
-        await server.WaitAssertion(() =>
-        {
-            var surgerySystem = entityManager.System<SharedSurgerySystem>();
-
-            var patient = entityManager.SpawnEntity("SurgeryTestPatient", mapData.GridCoords);
-
-            // No drape -> 1.0 (no penalty from drape layer)
-            Assert.That(surgerySystem.GetDrapeSpeedModifier(patient), Is.EqualTo(1f));
-
-            // Add SurgeryDrapedComponent with default (bedsheet improvised penalty)
-            entityManager.AddComponent<SurgeryDrapedComponent>(patient);
-            Assert.That(surgerySystem.GetDrapeSpeedModifier(patient), Is.EqualTo(1.5f));
-        });
-    }
-
-    /// <summary>
-    /// Verifies that GetDifficultyModifier returns the correct multiplier for each tier.
-    /// </summary>
-    [Test]
-    public async Task DifficultyModifierTest()
-    {
-        Assert.That(SharedSurgerySystem.GetDifficultyModifier(SurgeryDifficulty.Minor), Is.EqualTo(0.8f));
-        Assert.That(SharedSurgerySystem.GetDifficultyModifier(SurgeryDifficulty.Standard), Is.EqualTo(1.0f));
-        Assert.That(SharedSurgerySystem.GetDifficultyModifier(SurgeryDifficulty.Major), Is.EqualTo(1.3f));
-        Assert.That(SharedSurgerySystem.GetDifficultyModifier(SurgeryDifficulty.Critical), Is.EqualTo(1.5f));
     }
 
     /// <summary>
@@ -405,149 +249,6 @@ public sealed class SurgerySystemTest : GameTest
 
             // Clamping step also uses centralized default (2.0)
             Assert.That(SharedSurgerySystem.GetBaseStepDuration(majorProto.Steps[1]), Is.EqualTo(2.0f));
-        });
-    }
-
-    /// <summary>
-    /// Verifies that GetStepDuration correctly combines surface, drape, and difficulty modifiers.
-    /// </summary>
-    [Test]
-    public async Task StepDurationCombinesAllModifiersTest()
-    {
-        var server = Server;
-
-        var entityManager = server.ResolveDependency<IEntityManager>();
-        var protoManager = server.ResolveDependency<IPrototypeManager>();
-        var mapData = await Pair.CreateTestMap();
-
-        await server.WaitAssertion(() =>
-        {
-            var surgerySystem = entityManager.System<SharedSurgerySystem>();
-            var buckleSystem = entityManager.System<SharedBuckleSystem>();
-
-            protoManager.TryIndex<SurgeryProcedurePrototype>(TestProcedureId, out var proto);
-            var step = proto!.Steps[0]; // duration: 1.0
-
-            var patient = entityManager.SpawnEntity("SurgeryTestPatient", mapData.GridCoords);
-            var table = entityManager.SpawnEntity("SurgeryTestOperatingTable", mapData.GridCoords);
-
-            // Floor surgery, no drape, standard difficulty:
-            // 1.0 * 2.0 (surface) * 1.0 (no drape comp) * 1.0 (standard) = 2.0
-            Assert.That(surgerySystem.GetStepDuration(step, patient, SurgeryDifficulty.Standard),
-                Is.EqualTo(TimeSpan.FromSeconds(2.0)));
-
-            // Buckle to table, no drape, standard:
-            // 1.0 * 1.0 (surface) * 1.0 (no drape comp) * 1.0 (standard) = 1.0
-            var buckle = entityManager.GetComponent<BuckleComponent>(patient);
-            Assert.That(buckleSystem.TryBuckle(patient, patient, table, buckleComp: buckle), Is.True);
-            Assert.That(surgerySystem.GetStepDuration(step, patient, SurgeryDifficulty.Standard),
-                Is.EqualTo(TimeSpan.FromSeconds(1.0)));
-
-            // Add bedsheet drape (1.5x), major difficulty (1.3x):
-            // 1.0 * 1.0 (surface) * 1.5 (bedsheet) * 1.3 (major) = 1.95
-            entityManager.AddComponent<SurgeryDrapedComponent>(patient);
-            Assert.That(surgerySystem.GetStepDuration(step, patient, SurgeryDifficulty.Major).TotalSeconds,
-                Is.EqualTo(1.95).Within(0.001));
-
-            // Same bedsheet drape (1.5x), minor difficulty (0.8x):
-            // 1.0 * 1.0 (surface) * 1.5 (bedsheet) * 0.8 (minor) = 1.2
-            Assert.That(surgerySystem.GetStepDuration(step, patient, SurgeryDifficulty.Minor).TotalSeconds,
-                Is.EqualTo(1.2).Within(0.001));
-        });
-    }
-
-    /// <summary>
-    /// Verifies that the major test procedure loads with the correct difficulty.
-    /// </summary>
-    [Test]
-    public async Task ProcedureDifficultyLoadsTest()
-    {
-        var server = Server;
-        var protoManager = server.ResolveDependency<IPrototypeManager>();
-
-        await server.WaitAssertion(() =>
-        {
-            protoManager.TryIndex<SurgeryProcedurePrototype>(TestProcedureId, out var standard);
-            Assert.That(standard!.Difficulty, Is.EqualTo(SurgeryDifficulty.Standard));
-
-            protoManager.TryIndex<SurgeryProcedurePrototype>(TestProcedureMajorId, out var major);
-            Assert.That(major!.Difficulty, Is.EqualTo(SurgeryDifficulty.Major));
-        });
-    }
-
-    /// <summary>
-    /// Verifies that GetToolTierModifier returns the correct multiplier for each tier tag:
-    /// Experimental = 0.7, Advanced = 0.8, Standard = 1.0, no tag = 1.5 (improvised).
-    /// </summary>
-    [Test]
-    public async Task ToolTierModifierAllBranchesTest()
-    {
-        var server = Server;
-
-        var entityManager = server.ResolveDependency<IEntityManager>();
-        var mapData = await Pair.CreateTestMap();
-
-        await server.WaitAssertion(() =>
-        {
-            var surgerySystem = entityManager.System<SurgerySystem>();
-
-            var standard = entityManager.SpawnEntity("SurgeryTestScalpel", mapData.GridCoords);
-            var advanced = entityManager.SpawnEntity("SurgeryTestAdvancedScalpel", mapData.GridCoords);
-            var experimental = entityManager.SpawnEntity("SurgeryTestExperimentalScalpel", mapData.GridCoords);
-            var improvised = entityManager.SpawnEntity("SurgeryTestImprovisedTool", mapData.GridCoords);
-
-            Assert.That(surgerySystem.GetToolTierModifier(standard), Is.EqualTo(1.0f), "Standard tier");
-            Assert.That(surgerySystem.GetToolTierModifier(advanced), Is.EqualTo(0.8f), "Advanced tier");
-            Assert.That(surgerySystem.GetToolTierModifier(experimental), Is.EqualTo(0.7f), "Experimental tier");
-            Assert.That(surgerySystem.GetToolTierModifier(improvised), Is.EqualTo(1.5f), "Improvised (no tier tag)");
-        });
-    }
-
-    /// <summary>
-    /// Verifies that tool tier modifier integrates into the full duration calculation.
-    /// An advanced tool (0.8x) on the same step should produce a shorter DoAfter than standard (1.0x).
-    /// </summary>
-    [Test]
-    public async Task ToolTierAffectsStepDurationTest()
-    {
-        var server = Server;
-
-        var entityManager = server.ResolveDependency<IEntityManager>();
-        var protoManager = server.ResolveDependency<IPrototypeManager>();
-        var mapData = await Pair.CreateTestMap();
-
-        await server.WaitAssertion(() =>
-        {
-            var surgerySystem = entityManager.System<SurgerySystem>();
-            var buckleSystem = entityManager.System<SharedBuckleSystem>();
-
-            protoManager.TryIndex<SurgeryProcedurePrototype>(TestProcedureId, out var proto);
-            var step = proto!.Steps[0]; // duration: 1.0
-
-            var patient = entityManager.SpawnEntity("SurgeryTestPatient", mapData.GridCoords);
-            var table = entityManager.SpawnEntity("SurgeryTestOperatingTable", mapData.GridCoords);
-
-            // Buckle to table so surface = 1.0x, no drape so drape = 1.0x, standard difficulty
-            var buckle = entityManager.GetComponent<BuckleComponent>(patient);
-            Assert.That(buckleSystem.TryBuckle(patient, patient, table, buckleComp: buckle), Is.True);
-
-            // Base duration from GetStepDuration (without tool tier): 1.0 * 1.0 * 1.0 * 1.0 = 1.0
-            var baseDuration = (float) surgerySystem.GetStepDuration(step, patient, SurgeryDifficulty.Standard).TotalSeconds;
-
-            // Standard tool: 1.0 * 1.0 = 1.0
-            var standardTool = entityManager.SpawnEntity("SurgeryTestScalpel", mapData.GridCoords);
-            Assert.That(baseDuration * surgerySystem.GetToolTierModifier(standardTool),
-                Is.EqualTo(1.0f).Within(0.001f));
-
-            // Experimental tool: 1.0 * 0.7 = 0.7
-            var experimentalTool = entityManager.SpawnEntity("SurgeryTestExperimentalScalpel", mapData.GridCoords);
-            Assert.That(baseDuration * surgerySystem.GetToolTierModifier(experimentalTool),
-                Is.EqualTo(0.7f).Within(0.001f));
-
-            // Improvised tool: 1.0 * 1.5 = 1.5
-            var improvisedTool = entityManager.SpawnEntity("SurgeryTestImprovisedTool", mapData.GridCoords);
-            Assert.That(baseDuration * surgerySystem.GetToolTierModifier(improvisedTool),
-                Is.EqualTo(1.5f).Within(0.001f));
         });
     }
 
@@ -595,5 +296,4 @@ public sealed class SurgerySystemTest : GameTest
             }
         });
     }
-
 }
